+++ src/main/java/io/hhplus/reserve/common/annotation/Facade.java
@@ -1,4 +1,4 @@
-package io.hhplus.reserve.config.annotation;
+package io.hhplus.reserve.common.annotation;
 
 import org.springframework.core.annotation.AliasFor;
 import org.springframework.stereotype.Component;
